Move tokenize to base
Also add sleep to dataframe test so it passes on osx.
